+++ src/Symfony/Bundle/FrameworkBundle/Command/AboutCommand.php
@@ -29,15 +29,14 @@ use Symfony\Component\HttpKernel\KernelInterface;
  */
 class AboutCommand extends Command
 {
+    protected static $defaultName = 'about';
+
     /**
      * {@inheritdoc}
      */
     protected function configure()
     {
-        $this
-            ->setName('about')
-            ->setDescription('Displays information about the current project')
-        ;
+        $this->setDescription('Displays information about the current project');
     }
 
     /**