+++ src/Symfony/Bundle/FrameworkBundle/Command/TranslationDebugCommand.php
@@ -41,6 +41,8 @@ class TranslationDebugCommand extends Command
     const MESSAGE_UNUSED = 1;
     const MESSAGE_EQUALS_FALLBACK = 2;
 
+    protected static $defaultName = 'debug:translation';
+
     private $translator;
     private $loader;
     private $extractor;
@@ -60,7 +62,6 @@ class TranslationDebugCommand extends Command
     protected function configure()
     {
         $this
-            ->setName('debug:translation')
             ->setDefinition(array(
                 new InputArgument('locale', InputArgument::REQUIRED, 'The locale'),
                 new InputArgument('bundle', InputArgument::OPTIONAL, 'The bundle name or directory where to load the messages, defaults to app/Resources folder'),